tests: move gnrc_rpl_srh unittests to dedicated test application
