+++ tests/gnrc_rpl_srh/tests/01-run.py
@@ -330,6 +330,8 @@ def testfunc(child):
     global sniffer
     tap = get_bridge(os.environ["TAP"])
 
+    child.expect(r"OK \((\d+) tests\)")     # wait for and check result of unittests
+    print(".", end="", flush=True)
     lladdr_src = get_host_lladdr(tap)
     child.sendline("ifconfig")
     child.expect("HWaddr: (?P<hwaddr>[A-Fa-f:0-9]+)")
