--- tests/unittests/tests-gnrc_rpl_srh/Makefile
@@ -1,1 +1,0 @@
-include $(RIOTBASE)/Makefile.base
