--- tests/unittests/tests-gnrc_rpl_srh/Makefile.include
@@ -1,3 +1,0 @@
-USEMODULE += gnrc_ipv6
-USEMODULE += ipv6_addr
-USEMODULE += gnrc_rpl_srh
