--- tests/unittests/tests-gnrc_rpl_srh/tests-gnrc_rpl_srh.h
@@ -1,37 +1,0 @@
-/*
- * Copyright (C) 2016 Cenk Gündoğan <mail@cgundogan.de>
- *
- * This file is subject to the terms and conditions of the GNU Lesser
- * General Public License v2.1. See the file LICENSE in the top level
- * directory for more details.
- */
-
-/**
- * @addtogroup  unittests
- * @{
- *
- * @file
- * @brief       Unittests for the ``gnrc_rpl_srh`` module
- *
- * @author      Cenk Gündoğan <mail@cgundogan.de>
- */
-#ifndef TESTS_GNRC_RPL_SRH_H
-#define TESTS_GNRC_RPL_SRH_H
-
-#include "embUnit.h"
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-/**
- * @brief   The entry point of this test suite.
- */
-void tests_gnrc_rpl_srh(void);
-
-#ifdef __cplusplus
-}
-#endif
-
-#endif /* TESTS_GNRC_RPL_SRH_H */
-/** @} */
